AIs now wait till the castle is really completed before building anything. Especially with fast game speed it looked strange when many construction sites were placed and no serf left the castle.

+++ Freeserf.Core/AIStates/AIStateCastleBuilt.cs
@@ -19,12 +19,17 @@
  * along with freeserf.net. If not, see <http://www.gnu.org/licenses/>.
  */
 
+using System.Linq;
+
 namespace Freeserf.AIStates
 {
     class AIStateCastleBuilt : AIState
     {
         public override void Update(AI ai, Game game, Player player, PlayerInfo playerInfo, int tick)
         {
+            if (!game.GetPlayerBuildings(player, Building.Type.Castle).Any(c => c.IsDone()))
+                return;
+
             Kill(ai); // always kill before pushes!
 
             // enter idle
